trying to port some js files to ts
