+++ src/types/Context.ts
@@ -1,0 +1,5 @@
+
+/** The information available to the AI when calculating the scores. */
+interface Context {
+	
+}
