+++ src/types/FixedQualifier.ts
@@ -1,0 +1,18 @@
+/**
+  Returns a fixed score
+*/
+class FixedQualifier {
+	scorers: Scorer[];
+	action: Action;
+	value: number;
+	
+	constrcutor(scorers: Scorer[], action: Action, value: number) {
+		this.scorers = scorers;
+		this.action = action;
+		this.value = value;
+	}
+
+	score(context: Context) {
+		return Promise.resolve(this.value);
+	}
+}
