+++ src/types/Selector.ts
